# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Directorio de empresas - CCGAMT",
        "description": "Backend y API Acceso a datos.",
        "contact": {
            "email": "brayan-diaz3@upc.edu.co"
        },
        "license": {
            "name": "@Universidad Piloto de Colombia SAM - Todos los Derechos Reservados"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/roles": {
            "get": {
                "tags": [
                    "Roles"
                ],
                "summary": "Listar todos los roles",
                "operationId": "bf78d43aa8b5c576b78b08b73c404cb6",
                "responses": {
                    "200": {
                        "description": "Lista de roles"
                    }
                }
            },
            "post": {
                "tags": [
                    "Roles"
                ],
                "summary": "Crear un nuevo rol",
                "operationId": "19029466a195a478441990551cb1e8c9",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nombre"
                                ],
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "maxLength": 50
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Rol creado"
                    },
                    "400": {
                        "description": "Error de validación"
                    }
                }
            }
        },
        "/api/roles/{id}": {
            "get": {
                "tags": [
                    "Roles"
                ],
                "summary": "Obtener un rol por ID",
                "operationId": "bcca3423f39b15abd0d0214dd6169f3c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Rol encontrado"
                    },
                    "404": {
                        "description": "Rol no encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Roles"
                ],
                "summary": "Actualizar un rol existente",
                "operationId": "0444e2e7859bbaa81749d090c06415a1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nombre"
                                ],
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "maxLength": 50
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Rol actualizado"
                    },
                    "400": {
                        "description": "Error de validación"
                    },
                    "404": {
                        "description": "Rol no encontrado"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Roles"
                ],
                "summary": "Eliminar un rol",
                "operationId": "de55f66883ee7b0d98d50e5e82e94f7e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Rol eliminado"
                    },
                    "404": {
                        "description": "Rol no encontrado"
                    }
                }
            }
        },
...
        "/api/usuarios": {
            "get": {
                "tags": [
                    "Usuarios"
                ],
                "summary": "Lista todos los usuarios",
                "operationId": "e9d4d8b419dbef04b5748fbf9ccf3881",
                "responses": {
                    "200": {
                        "description": "Lista de usuarios"
                    }
                }
            },
            "post": {
                "tags": [
                    "Usuarios"
                ],
                "summary": "Crear un usuario",
                "operationId": "4e9dbcbdaaaf9faab196a82d5cbaea23",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "example": "Juan Pérez"
                                    },
                                    "correo": {
                                        "type": "string",
                                        "example": "juan@example.com"
                                    },
                                    "contrasena": {
                                        "type": "string",
                                        "example": "password123"
                                    },
                                    "id_rol": {
                                        "type": "integer",
                                        "example": 2
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Usuario creado exitosamente"
                    }
                }
            }
        },
        "/api/usuarios/{id}": {
            "get": {
                "tags": [
                    "Usuarios"
                ],
                "summary": "Obtener un usuario por ID",
                "operationId": "e9ac05ddc4ea46d8848b012d7d92e3ac",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalles del usuario"
                    }
                }
            },
            "put": {
                "tags": [
                    "Usuarios"
                ],
                "summary": "Actualizar un usuario",
                "operationId": "27ac608b6dd8b8e7c7c45b10c5210b3e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "example": "Juan Pérez"
                                    },
                                    "correo": {
                                        "type": "string",
                                        "example": "juan@example.com"
                                    },
                                    "contrasena": {
                                        "type": "string",
                                        "example": "newpassword123"
                                    },
                                    "id_rol": {
                                        "type": "integer",
                                        "example": 2
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Usuario actualizado correctamente"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Usuarios"
                ],
                "summary": "Eliminar un usuario",
                "operationId": "ea700764859a665afd38bb39c2842f1c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Usuario eliminado correctamente"
                    }
                }
            }
        }
    },
...
}